Reorganize publishing into convention plugin
