+++ buildSrc/src/main/kotlin/Ci.kt
@@ -1,7 +1,7 @@
 object Ci {
    // this is the version used for building snapshots
    // .GITHUB_RUN_NUMBER-snapshot will be appended
-   private const val snapshotBase = "1.0.0"
+   private const val snapshotBase = "2.0.0"
 
    private val githubRunNumber = System.getenv("GITHUB_RUN_NUMBER")
