+++ buildSrc/src/main/kotlin/kotest-publishing-conventions.gradle.kts
@@ -1,0 +1,87 @@
+import org.gradle.api.publish.PublicationContainer
+import org.gradle.api.publish.PublishingExtension
+import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.kotlin.dsl.get
+import org.gradle.kotlin.dsl.invoke
+import org.gradle.kotlin.dsl.provideDelegate
+import org.gradle.kotlin.dsl.register
+import org.gradle.kotlin.dsl.repositories
+import org.gradle.kotlin.dsl.signing
+
+plugins {
+   signing
+   `java-library`
+   `maven-publish`
+}
+
+group = "io.kotest.extensions"
+version = Ci.version
+
+val publications: PublicationContainer = (extensions.getByName("publishing") as PublishingExtension).publications
+
+val ossrhUsername: String by project
+val ossrhPassword: String by project
+val signingKey: String? by project
+val signingPassword: String? by project
+
+signing {
+   useGpgCmd()
+   if (signingKey != null && signingPassword != null) {
+      useInMemoryPgpKeys(signingKey, signingPassword)
+   }
+   if (Ci.isRelease) {
+      sign(publications)
+   }
+}
+
+java {
+   withJavadocJar()
+   withSourcesJar()
+}
+
+publishing {
+   repositories {
+      maven {
+         val releasesRepoUrl = uri("https://oss.sonatype.org/service/local/staging/deploy/maven2/")
+         val snapshotsRepoUrl = uri("https://oss.sonatype.org/content/repositories/snapshots/")
+         name = "deploy"
+         url = if (Ci.isRelease) releasesRepoUrl else snapshotsRepoUrl
+         credentials {
+            username = System.getenv("OSSRH_USERNAME") ?: ossrhUsername
+            password = System.getenv("OSSRH_PASSWORD") ?: ossrhPassword
+         }
+      }
+   }
+
+   publications {
+      register("mavenJava", MavenPublication::class) {
+         from(components["java"])
+         pom {
+            name.set("kotest-extensions-embedded-kafka")
+            description.set("Kotest extension for embedded kafka")
+            url.set("http://www.github.com/kotest/kotest-extensions-embedded-kafka")
+
+            scm {
+               connection.set("scm:git:http://www.github.com/kotest/kotest-extensions-embedded-kafka")
+               developerConnection.set("scm:git:http://github.com/sksamuel")
+               url.set("http://www.github.com/kotest/kotest-extensions-embedded-kafka")
+            }
+
+            licenses {
+               license {
+                  name.set("The Apache 2.0 License")
+                  url.set("https://opensource.org/licenses/Apache-2.0")
+               }
+            }
+
+            developers {
+               developer {
+                  id.set("sksamuel")
+                  name.set("Stephen Samuel")
+                  email.set("sam@sksamuel.com")
+               }
+            }
+         }
+      }
+   }
+}
